# spring/docs/openapi.yaml
openapi: "3.0.1"
info:
  title: "Spring部分的Swagger文档"
  description: "这是项目的Spring部分的Swagger文档"
  version: "1.0.0"
...
paths:
...
  /collects:
    post:
      tags:
      - "文章收藏模块"
      summary: "添加收藏"
      description: "为文章添加收藏"
      operationId: "addCollect"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleCollectDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
    delete:
      tags:
      - "文章收藏模块"
      summary: "取消收藏"
      description: "取消对文章的收藏"
      operationId: "removeCollect"
      parameters:
      - name: "article_id"
        in: "query"
        description: "文章ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "user_id"
        in: "query"
        description: "用户ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
  /collects/counts/batch:
    post:
      tags:
      - "文章收藏模块"
      summary: "批量查询收藏数（内部）"
      description: "根据文章ID列表批量查询收藏数，供内部服务远程调用"
      operationId: "getCollectCountsByArticleIds"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchIdsDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultBatchCountVO"
...
  /collects/user/{user_id}:
    get:
      tags:
      - "文章收藏模块"
      summary: "查询用户的所有收藏"
      description: "分页查询某个用户的所有收藏记录（包含文章详情）"
      operationId: "listUserCollects"
      parameters:
      - name: "user_id"
        in: "path"
        description: "用户ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "page"
        in: "query"
        description: "页码"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      - name: "size"
        in: "query"
        description: "每页数量"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOArticleCollectVO"
...
  /collects/statistics/monthly-trend/{userId}:
    get:
      tags:
      - "文章收藏模块"
      summary: "获取用户月度收藏趋势（内部）"
      description: "获取用户本月收藏的趋势，供内部服务远程调用"
      operationId: "getMonthlyCollectTrend"
      parameters:
      - name: "user_id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultMapDataVO"
...
  /collects/neo4j-sync:
    get:
      tags:
      - "文章收藏模块"
      summary: "获取收藏表数据用于Neo4j同步（内部）"
      description: "获取收藏表数据，支持增量同步，供FastAPI同步Neo4j使用"
      operationId: "getNeo4jSyncCollects"
      parameters:
      - name: "updated_after"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListArticleRelationSyncVO"
  /collects/count/{article_id}:
    get:
      tags:
      - "文章收藏模块"
      summary: "获取文章的收藏数"
      description: "获取某篇文章的总收藏数"
      operationId: "getCollectCount"
      parameters:
      - name: "article_id"
        in: "path"
        description: "文章ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultCollectCountVO"
  /collects/check:
    get:
      tags:
      - "文章收藏模块"
      summary: "检查用户是否收藏"
      description: "查询用户是否收藏过某篇文章"
      operationId: "isCollected"
      parameters:
      - name: "article_id"
        in: "query"
        description: "文章ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "user_id"
        in: "query"
        description: "用户ID"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultCollectCheckVO"
...
  /articles/user/{id}:
    get:
      tags:
      - "文章模块"
      summary: "获取用户文章"
      description: "返回用户文章，可指定是否只查询已发布的文章"
      operationId: "getArticlesByUserId"
      parameters:
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      - name: "size"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "published"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 0
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOArticleWithCategoryVO"
...
  /articles/neo4j-sync:
    get:
      tags:
      - "文章模块"
      summary: "获取文章表数据用于Neo4j同步（内部）"
      description: "获取文章表数据，支持增量同步，供FastAPI同步Neo4j使用"
      operationId: "getNeo4jSyncArticles"
      parameters:
      - name: "updated_after"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListArticleSyncVO"
...
components:
  schemas:
...
    ResultVoid:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "null"
          description: "返回数据"
          nullable: true
      description: "统一响应"
...
    BatchIdsDTO:
      required:
      - "ids"
      type: "object"
      properties:
        ids:
          type: "array"
          description: "ID列表"
          items:
            type: "integer"
            description: "ID列表"
            format: "int64"
      description: "批量ID请求"
...
    BatchCountVO:
      type: "object"
      properties:
        counts:
          type: "array"
          description: "ID 与数量列表"
          items:
            $ref: "#/components/schemas/IdCountVO"
      description: "批量计数结果"
    IdCountVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "ID"
          format: "int64"
        count:
          type: "integer"
          description: "数量"
          format: "int64"
      description: "ID 与数量"
    ResultBatchCountVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/BatchCountVO"
      description: "统一响应"
...
    ArticleCollectDTO:
      required:
      - "article_id"
      - "user_id"
      type: "object"
      properties:
        article_id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        user_id:
          type: "integer"
          description: "用户ID"
          format: "int64"
      description: "文章收藏请求"
...
    ArticleWithCategoryVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        title:
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章内容"
        user_id:
          type: "integer"
          description: "作者用户ID"
          format: "int64"
        username:
          type: "string"
          description: "作者名称"
        tags:
          type: "string"
          description: "标签，多个用英文逗号分隔"
        status:
          type: "integer"
          description: "状态，0草稿 1已发布"
          format: "int32"
        views:
          type: "integer"
          description: "浏览量"
          format: "int32"
        sub_category_id:
          type: "integer"
          description: "子分类ID"
          format: "int32"
        sub_category_name:
          type: "string"
          description: "子分类名称"
        category_id:
          type: "integer"
          description: "父分类ID"
          format: "int64"
        category_name:
          type: "string"
          description: "父分类名称"
        create_at:
          type: "string"
          description: "创建时间"
          format: "date-time"
        update_at:
          type: "string"
          description: "更新时间"
          format: "date-time"
      description: "带分类的文章"
...
    MapDataVO:
      type: "object"
      properties:
        data:
          type: "object"
          additionalProperties:
            type: "object"
            description: "动态键值对"
          description: "动态键值对"
      description: "键值数据"
    ResultMapDataVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/MapDataVO"
      description: "统一响应"
...
    ArticleRelationSyncVO:
      type: "object"
      properties:
        user_id:
          type: "integer"
          description: "用户ID"
          format: "int64"
        article_id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        created_time:
          type: "string"
          description: "关系创建时间"
          format: "date-time"
      description: "用户文章关系同步数据"
    ResultListArticleRelationSyncVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/ArticleRelationSyncVO"
      description: "统一响应"
...
    ArticleCollectVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "收藏ID"
          format: "int64"
        article_id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        title:
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章内容"
        user_id:
          type: "integer"
          description: "作者用户ID"
          format: "int64"
        author_name:
          type: "string"
          description: "作者名称"
        tags:
          type: "string"
          description: "标签，多个用英文逗号分隔"
        status:
          type: "integer"
          description: "状态，0草稿 1已发布"
          format: "int32"
        views:
          type: "integer"
          description: "浏览量"
          format: "int32"
        sub_category_id:
          type: "integer"
          description: "子分类ID"
          format: "int32"
        category_name:
          type: "string"
          description: "父分类名称"
        sub_category_name:
          type: "string"
          description: "子分类名称"
        article_create_at:
          type: "string"
          description: "文章创建时间"
          format: "date-time"
        article_update_at:
          type: "string"
          description: "文章更新时间"
          format: "date-time"
        collected_time:
          type: "string"
          description: "收藏时间"
          format: "date-time"
      description: "文章收藏视图"
    PageVOArticleCollectVO:
      type: "object"
      properties:
        total:
          type: "integer"
          description: "总记录数"
          format: "int64"
        list:
          type: "array"
          description: "当前页数据"
          items:
            $ref: "#/components/schemas/ArticleCollectVO"
      description: "分页结果"
    ResultPageVOArticleCollectVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/PageVOArticleCollectVO"
      description: "统一响应"
    CollectCountVO:
      type: "object"
      properties:
        collect_count:
          type: "integer"
          description: "收藏数"
          format: "int64"
      description: "收藏数量"
    ResultCollectCountVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/CollectCountVO"
      description: "统一响应"
    CollectCheckVO:
      type: "object"
      properties:
        collected:
          type: "boolean"
          description: "是否已收藏"
      description: "收藏状态"
    ResultCollectCheckVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/CollectCheckVO"
      description: "统一响应"
...
    PageVOArticleWithCategoryVO:
      type: "object"
      properties:
        total:
          type: "integer"
          description: "总记录数"
          format: "int64"
        list:
          type: "array"
          description: "当前页数据"
          items:
            $ref: "#/components/schemas/ArticleWithCategoryVO"
      description: "分页结果"
    ResultPageVOArticleWithCategoryVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/PageVOArticleWithCategoryVO"
      description: "统一响应"
...
    ArticleSyncVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        title:
          type: "string"
          description: "文章标题"
        content:
          type: "string"
          description: "文章正文"
        user_id:
          type: "integer"
          description: "作者用户ID"
          format: "int64"
        tags:
          type: "string"
          description: "标签，多个标签以逗号分隔"
        status:
          type: "integer"
          description: "文章状态，0 草稿，1 已发布"
          format: "int32"
        views:
          type: "integer"
          description: "浏览量"
          format: "int32"
        sub_category_id:
          type: "integer"
          description: "子分类ID"
          format: "int32"
        create_at:
          type: "string"
          description: "创建时间"
          format: "date-time"
        update_at:
          type: "string"
          description: "更新时间"
          format: "date-time"
      description: "文章同步数据"
    ResultListArticleSyncVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/ArticleSyncVO"
      description: "统一响应"
...